server/requirements.txt
Django==3.0.8
psycopg2-binary==2.8.5 # работа с базой данных postgres

dj-database-url==0.5.0 # передача настроек БД из env окружения
gunicorn==20.0.4 # http сервер, который свзявает nginx с django приложением

ipdb==0.13.3 # отладка программ
ipython==7.12.0 # приятная работа в shell
django-debug-toolbar==2.2 # панель для отладки запросов в бд

Pillow==7.2.0 # для работы с изображениями
djangorestframework==3.11.0 # API
django-url-filter==0.3.15 # улучшенные фильтры для DRF
bs4==0.0.1 # BeautifulSoup
drf-yasg==1.17.1 # Swagger
django-cors-headers==3.4.0
